Fix - Error floating precision

+++ src/color/color.h
@@ -10,7 +10,7 @@
 
 using color = vec3;
 
-inline double linear_to_gamma(double linear_component) {
+inline float linear_to_gamma(float linear_component) {
     if (linear_component > 0) {
         return sqrt(linear_component);
     }
