+++ src/hittable/hitabble.h
@@ -14,10 +14,10 @@ public:
     point3  p;
     vec3 normal;
     shared_ptr<material> mat;
-    double t;
+    float t;
     bool front_face;
-    double u;
-    double v;
+    float u;
+    float v;
 
     void set_face_normal(const ray& r, const vec3& outward_normal) {
         // Sets the hit record normal vector.
