+++ src/interval/interval.h
@@ -9,31 +9,31 @@
 
 class interval {
 public:
-    double min, max;
+    float min, max;
 
     interval() : min(+infinity), max(-infinity) {} // Default interval is empty
 
-    interval(double min, double max) : min(min), max(max) {}
+    interval(float min, float max) : min(min), max(max) {}
 
-    double size() const {
+    float size() const {
         return max-min;
     }
 
-    bool contains(double x) const {
+    bool contains(float x) const {
         return min <= x && x <= max;
     }
 
-    bool surrounds(double x) const {
+    bool surrounds(float x) const {
         return min < x && x < max;
     }
     
-    double clamp(double x) const {
+    float clamp(float x) const {
         if (x < min) return min;
         if (x > max) return max;
         return x;
     }
 
-    [[nodiscard]] interval expand(double delta) const {
+    [[nodiscard]] interval expand(float delta) const {
         auto padding = delta/2;
         return {min - padding, max + padding};
     }
