+++ src/texture/texture.h
@@ -13,7 +13,7 @@ class texture {
 public:
     virtual ~texture() = default;
 
-    virtual color value(double u, double v, const point3& p) const = 0;
+    virtual color value(float u, float v, const point3& p) const = 0;
 };
 
 class solid_color : public texture {
@@ -22,9 +22,9 @@ private:
 public:
     solid_color(const color& albedo) : albedo(albedo) {}
 
-    solid_color(double red, double green, double blue) : solid_color(color(red,green,blue)) {}
+    solid_color(float red, float green, float blue) : solid_color(color(red,green,blue)) {}
 
-    color value(double u, double v, const point3& p) const override {
+    color value(float u, float v, const point3& p) const override {
         return albedo;
     }
 
@@ -32,16 +32,16 @@ public:
 
 class checker_texture : public texture {
 public:
-    checker_texture(double scale, shared_ptr<texture> even, shared_ptr<texture> odd)
+    checker_texture(float scale, shared_ptr<texture> even, shared_ptr<texture> odd)
             : inv_scale(1.0 / scale), even(even), odd(odd) {}
 
-    checker_texture(double scale, const color& c1, const color& c2)
+    checker_texture(float scale, const color& c1, const color& c2)
             : inv_scale(1.0 / scale),
               even(make_shared<solid_color>(c1)),
               odd(make_shared<solid_color>(c2))
     {}
 
-    color value(double u, double v, const point3& p) const override {
+    color value(float u, float v, const point3& p) const override {
         auto xInteger = int(std::floor(inv_scale * p.x()));
         auto yInteger = int(std::floor(inv_scale * p.y()));
         auto zInteger = int(std::floor(inv_scale * p.z()));
@@ -52,21 +52,21 @@ public:
     }
 
 private:
-    double inv_scale;
+    float inv_scale;
     shared_ptr<texture> even;
     shared_ptr<texture> odd;
 };
 
 class noise_texture : public texture {
 private:
-    double scale;
+    float scale;
     perlin noise;
 public:
     noise_texture() {}
 
-    noise_texture(double scale) : scale(scale) {}
+    noise_texture(float scale) : scale(scale) {}
 
-    color value(double u, double v, const point3& p) const override {
+    color value(float u, float v, const point3& p) const override {
         return color(.5, .5, .5) * (1 + sin(scale * p.z() + 10 * noise.turb(p, 7)));
     }
